Refactor BuildingCardFactory
