+++ packages/backend/domain/src/main/java/tw/waterballsa/gaas/citadels/domain/CitadelsGame.java
@@ -1,8 +1,6 @@
 package tw.waterballsa.gaas.citadels.domain;
 
 import tw.waterballsa.gaas.citadels.domain.BuildingCard.BuildingCard;
-import tw.waterballsa.gaas.citadels.domain.BuildingCard.BuildingCardFactory;
-
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
@@ -66,6 +64,6 @@ public class CitadelsGame {
     }
 
     public List<BuildingCard> getTwoCards() {
-        return Arrays.asList(new BuildingCard("test", 3, BuildingCardFactory.Color.BLUE), new BuildingCard("test", 2, BuildingCardFactory.Color.BLUE));
+        return Arrays.asList(new BuildingCard("test", 3, BuildingCard.Color.BLUE), new BuildingCard("test", 2, BuildingCard.Color.BLUE));
     }
 }
